#OMJF-12345 【接口劣化】<fix> 修复请求数int溢出导致显示负数
- currentCount/previousCount 改为 long 类型
- 去除 (int) 强转，防止大数值溢出

+++ backend/src/main/java/com/ykc/hubble/service/GatewayService.java
@@ -298,8 +298,8 @@ public class GatewayService {
             vo.setCurrentAvgTime(Math.round(currentAvg * 10.0) / 10.0);
             vo.setPreviousAvgTime(Math.round(previousAvg * 10.0) / 10.0);
             vo.setDegradationRate(Math.round(rate * 10.0) / 10.0);
-            vo.setCurrentCount((int) currentCount);
-            vo.setPreviousCount((int) previousCount);
+            vo.setCurrentCount(currentCount);
+            vo.setPreviousCount(previousCount);
             result.add(vo);
         }
 
@@ -335,8 +335,8 @@ public class GatewayService {
                 vo.setCurrentAvgTime(Math.round(currentAvg * 10.0) / 10.0);
                 vo.setPreviousAvgTime(Math.round(previousAvg * 10.0) / 10.0);
                 vo.setDegradationRate(Math.round(rate * 10.0) / 10.0);
-                vo.setCurrentCount((int) currentCount);
-                vo.setPreviousCount((int) previousCount);
+                vo.setCurrentCount(currentCount);
+                vo.setPreviousCount(previousCount);
                 result.add(vo);
             }
             result.sort((a, b) -> Double.compare(b.getDegradationRate(), a.getDegradationRate()));
